requirements/base.txt
django==1.3.1
django-tastypie==0.9.11
django-mongonaut==0.2.10
mongoengine==0.5.2
pymongo==2.1.1
